Updated Ratingrepository to now also support a PATCH operation to modify the rating. To reset a given rating to default we expect "ratingModifier": 0. This changes it from the prev rating back to the default value. If the ratingModifier is the same nothing happens. It the modifier differs we compensate the old modification and then apply the new one.

+++ app/util/JsonKeys.java
@@ -52,8 +52,8 @@ public class JsonKeys {
     public static final String USER_PASSWORD = "password";
     public static final String USER_EMAIL = "email";
     public static final String USER_GROUPS = "groups";
-    public static final int USER_PASSWORD_MIN_LENGTH=5;
-    public static final int USER_NAME_MIN_LENGTH=5;
+    public static final long USER_PASSWORD_MIN_LENGTH=5;
+    public static final long USER_NAME_MIN_LENGTH = 5;
     public static final String USER_JSON_ELEMENTS = USER_NAME + ", " + USER_EMAIL + ", " + USER_PASSWORD + ", " + USER_GROUPS;
     //Tag
     public static final String TAG_ID = "tagId";
@@ -111,9 +111,6 @@ public class JsonKeys {
     //Media Upload
     public static final String MEDIA_ID = "mediaId";
     public static final String MEDIA_TYPE = "mediaType";
-
-
-
 
 
 }
